item controller crud completed

+++ 14_Spring_Data_JPA/src/main/java/lk/ijse/spring/service/impl/ItemServiceImpl.java
@@ -24,17 +24,21 @@ public class ItemServiceImpl implements ItemService {
     @Autowired
     ModelMapper mapper;
 
-    public ItemServiceImpl() {
-        System.out.println("ItemServiceImpl Instantiated");
-    }
-
     @Override
-    public void addItem(ItemDTO item) {
-        itemRepo.save(mapper.map(item, Item.class));
+    public void addItem(ItemDTO dto) {
+        if (itemRepo.existsById(dto.getCode())) {
+            throw new RuntimeException(dto.getCode()+" is already available, please insert a new ItemID");
+        }
+
+        Item map = mapper.map(dto, Item.class);
+        itemRepo.save(map);
     }
 
     @Override
     public void deleteItem(String code) {
+        if (!itemRepo.existsById(code)) {
+            throw new RuntimeException(code+ " This Item is not available, please check the ItemID before delete.!");
+        }
         itemRepo.deleteById(code);
     }
 
@@ -43,18 +47,22 @@ public class ItemServiceImpl implements ItemService {
         List<Item> all = itemRepo.findAll();
         return mapper.map(all, new TypeToken<List<ItemDTO>>() {
         }.getType());
-        //new TypeToken<>(){}.getType()
-        //new TypeToken<List<CustomerDTO>>(){}.getType()
     }
 
     @Override
     public ItemDTO findItem(String code) {
+        if (!itemRepo.existsById(code)) {
+            throw new RuntimeException(code+ " This Item is not available, please check the ItemID.!");
+        }
         Item item = itemRepo.findById(code).get();
         return mapper.map(item,ItemDTO.class);
     }
 
     @Override
     public void updateItem(ItemDTO i) {
+        if (!itemRepo.existsById(i.getCode())) {
+            throw new RuntimeException(i.getCode()+ " This Item is not available, please check the ItemID before update.!");
+        }
         Item map = mapper.map(i, Item.class);
         itemRepo.save(map);
     }
